requirements.txt
echo fastapi>=0.95.0
uvicorn>=0.21.1
scikit-learn>=1.2.0
pandas>=1.5.0
python-multipart
fastapi-cors>=0.1.0
joblib>=1.2.0